fix 32/64 compat + fix ULA-OP scan
Former-commit-id: 51fc3334cccadc564622b06c744ec5a9892d5148 [formerly 8a4a16d2ae7de406d97ea5df11339bb02b08e15b]
Former-commit-id: aa15469456294f7dbe0ee407431d34bb72ec6425

+++ src/Acquisition.hpp
@@ -31,7 +31,7 @@ class SRC_API Measure
     //    Measurement m_measurement;
     //    virtual SensorSpec::Format getFormat() = 0;
     const unsigned char* const m_data;
-    const size_t m_size;
+    const uint64_t m_size; // compatibility 32/64 bits
 
     bool operator==( const Measure& measure ) const;
     bool operator!=( const Measure& measure ) const;
